+++ littletree/__init__.py
@@ -2,12 +2,14 @@ from .basenode import BaseNode
 from .maxdepth import MaxDepth
 from .node import Node
 from .nodemixin import NodeMixin
+from .route import Route
 
 __all__ = [
     "BaseNode",
     "Node",
     "NodeMixin",
     "MaxDepth",
+    "Route",
 ]
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"