+++ littletree/exceptions.py
@@ -1,4 +1,8 @@
-class DuplicateParentError(Exception):
+class TreeException(Exception):
+    pass
+
+
+class DuplicateParentError(TreeException):
     def __init__(self, node):
         self.node = node
 
@@ -7,7 +11,7 @@ class DuplicateParentError(Exception):
                 f"Consider calling node.copy() or node.detach() first.")
 
 
-class DuplicateChildError(Exception):
+class DuplicateChildError(TreeException):
     def __init__(self, node, identifier):
         self.node = node
         self.identifier = identifier
@@ -16,7 +20,7 @@ class DuplicateChildError(Exception):
         return f"{self.identifier} is already present in {self.node}"
 
 
-class LoopError(Exception):
+class LoopError(TreeException):
     def __init__(self, node, ancestor):
         self.node = node
         self.ancestor = ancestor
@@ -24,5 +28,11 @@ class LoopError(Exception):
     def __str__(self):
         return f"{self.node} can't become parent of {self.ancestor}"
 
-class DifferentTreeError(Exception):
-    pass
+
+class DifferentTreeError(TreeException):
+    def __init__(self, root1, root2):
+        self.root1 = root1
+        self.root2 = root2
+
+    def __str__(self):
+        return f"Nodes have different roots: {self.root1} vs {self.root2}"